Fixed skipping windows too quickly
If the button to manually switch to a new window had been pressed, the
next window was replaced by the one after that too early because the
value indicating how long it was in memory was not reset. Now this value
is reset and the delay will be the desired delay.