Communication from view directly to controller
User interaction events were previously communicated from the View
object to the Incapy object (the overall interface towards the user) and
only from there to the controller. This additional step is unnecessary
and is thus removed in this commit, making code cleaner. With this the
parsing of messages got a lot more general, as the Incapy class does no
longer need to know how to process every single message.
Also the communication was split up in two parts: Changed values and
events, for which separate methods are introduced in this commit.